requirements.txt
pywhatkit>=5.0
pandas
python-dotenv
schedule
